1_foundations/requirements.txt
requests>=2.31.0
python-dotenv>=0.19.0
gradio>=4.0.0
pypdf>=3.0.0
openai>=1.0.0
openai-agents
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
tiktoken>=0.5.1
faiss-cpu>=1.7.4
numpy>=1.24.0